// frontend/src/app/core/api/generated/model/inventory-response.ts
/**
 * Inventory API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InventoryResponse { 
    warehouseId?: string;
    productId?: string;
    quantity?: number;
    reservedQuantity?: number;
    availableQuantity?: number;
    updatedAt?: string;
}